# ci Don't always re-install dependencies, which speeds up successive builds