--- a/tox.ini
+++ b/tox.ini
@@ -1,19 +1,31 @@
 [tox]
 env_list =
-    node, eslint, stylelint, pylint, py312
+    pytest
+    pylint
+    eslint
+    stylelint
 minversion = 4.6.0
 skip_missing_interpreters = false
 
 [testenv]
-description = run the tests with pytest
+description = base environment
 package = wheel
 wheel_build_env = .pkg
 deps =
+    pytest: -r requirements.txt
+    eslint,stylelint: nodeenv
+commands_pre =
+    {es,style}lint: ./ci/node-install.sh
+allowlist_externals =
+    {es,style}lint: ./ci/node-install.sh
+
+[testenv:pytest]
+description = run the tests with pytest
+deps =
+    {[testenv]deps}
     pytest>=6
     pytest-mock>=3
-    -rrequirements.txt
-commands =
-    pytest {tty:--color=yes} {posargs}
+commands = pytest {tty:--color=yes} {posargs}
 
 [testenv:pylint]
 description = run python black linters
@@ -22,39 +34,12 @@ deps =
     black==24.8.0
 commands = black --check {posargs: ./tests ./dashboard}
 
-[testenv:node]
-description = setup Node.js environment
-skip_install = true
-deps =
-    nodeenv
-commands =
-    nodeenv --prebuilt -p --node 18.20.4
-    npm install -g npm@latest
-allowlist_externals =
-    node
-    npm
-    npx
-
 [testenv:eslint]
 description = run eslint
 skip_install = true
-deps =
-    {[testenv:node]deps}
-commands =
-    {[testenv:node]commands}
-    npm ci
-    npm run lint:js
-depends = node
-allowlist_externals = npm
+commands = npm run lint:js
 
 [testenv:stylelint]
 description = run stylelint
 skip_install = true
-deps =
-    {[testenv:node]deps}
-commands =
-    {[testenv:node]commands}
-    npm ci
-    npm run lint:css
-depends = node
-allowlist_externals = npm
+commands = npm run lint:css