--- a/tox.ini
+++ b/tox.ini
@@ -11,6 +11,9 @@ skip_missing_interpreters = false
 description = base environment
 package = wheel
 wheel_build_env = .pkg
+set_env =
+    NODE_VERSION=18.20.4
+    NPM_VERSION=10.9.0
 deps =
     pytest: -r requirements.txt
     eslint,stylelint: nodeenv