(Sample size: 1)

Now (some output ommitted):

    # Do a clean build
    rm -rf .tox; time tox
    tox  17.25s user 8.84s system 49% cpu 52.303 total
    # Do a successive build
    time tox
    tox  1.72s user 0.38s system 86% cpu 2.436 total

Before (some output ommitted):

    # Do a clean build
    rm -rf .tox; time tox
    tox  19.69s user 11.41s system 46% cpu 1:06.32 total
    # Do a successive build
    time tox
    tox  3.57s user 3.64s system 24% cpu 29.537 total

It can go _faster_, but that'd come at the cost of having tox managing our dependencies when they change, so I opted not to do that. Though, we should keep that in our back pocket if we get to that point.